Implement toString( std::pair<T1,T2> )
- Transplanted the earlier implementation of StringMaker<pair>from
  ToStringPair.cpp to catch_tostring.h

 - Remove the specialisation of toString( pair<int,int> ) from
   TrickyTests.cpp
