+++ include/internal/catch_tostring.h
@@ -237,12 +237,27 @@ struct StringMaker<std::tuple<Types...>> {
 };
 #endif
 
+template<typename T1, typename T2>
+struct StringMaker<std::pair<T1,T2> > {
+    static std::string convert( const std::pair<T1,T2>& pair ) {
+        std::ostringstream oss;
+        oss << "{ "
+            << Catch::toString( pair.first )
+            << ", "
+            << Catch::toString( pair.second )
+            << " }";
+        return oss.str();
+    }
+};
+
+
 namespace Detail {
     template<typename T>
     std::string makeString( T const& value ) {
         return StringMaker<T>::convert( value );
     }
 } // end namespace Detail
+
 
 /// \brief converts any type to a string
 ///
